requirements.txt
# Evidence Finder + Data Builder — pilot
pandas>=2.0.0
openpyxl>=3.1.0
python-docx>=1.0.0
pyyaml>=6.0
requests>=2.28.0
beautifulsoup4>=4.12.0
selenium>=4.15.0

# UI (choose one or both)
streamlit>=1.28.0
flask>=3.0.0

# Tests
pytest>=7.0.0
